object_detection/requirements.txt
# Django framework
Django>=5.1

# REST API
djangorestframework

# CORS headers
django-cors-headers

# Image handling
numpy
opencv-python-headless
Pillow

# Machine learning / YOLO
ultralytics
torch  # CPU version; change if using GPU

# Production server
gunicorn
